Retain column information in parsing.
We previously were not storing column information when parsing source
files which makes it much harder to correlate input to output. This
patch initializes already existing data structures with this information
when available. Locations are emitted as `<line>:<character>` so that
e.g., the second character on the fourth line would be displayed as
`4:2`. This extends to ranges as well which look e.g., like `1:1-2:30`.